# openapi.yaml
# Docs: https://swagger.io/docs/specification/about/
---
openapi: 3.0.3
info:
  title: Cert Manager Selfservice API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /info:
    get:
      summary: Returns information about the cert-manager-selfservice
      responses:
        '200':
          description: A json object represent the info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Info'
  /status:
    get:
      summary: Returns status of the cert-manager-selfservice
      responses:
        '200':
          description: A json object represent the status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /certificate/{domain}:
    get:
      summary: Return and eventually create certificate for the given domain
      parameters:
        - name: domain
          in: path
          description: The domain to get the certificate for
          required: true
          schema:
            type: string
        - name: format
          in: query
          description: Format of the result, defaults to json
          required: false
          schema:
            type: string
            default: json
            enum:
              - json
              - crt
              - key
              - ca
      responses:
        '200':
          description: A json object represent the certificate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Certificate'
            html/text: {}
        '202':
          description: Server tries to aquire the certificate you should try again later
          content:
            html/text: {}
        '404':
          description: The cert cannot be found it also canot be created
          content:
            html/text: {}
components:
  schemas:
    Info:
      type: object
      required:
        - version
      properties:
        version:
          description: The version of the cert-manager-selfservice
          type: string
    Status:
      type: object
      properties:
        messages:
          description: This are event messages produced by the cert-manager-selfservice
          type: array
          items:
            $ref: '#/components/schemas/StatusMessage'
        certificates:
          $ref: '#/components/schemas/StatusCertificates'
      required:
        - messages
        - certificates
    StatusMessage:
      type: object
      required:
        - time
        - message
        - severity
      properties:
        time:
          type: string
          format: date-time
        severity:
          type: string
          enum:
            - info
            - warning
            - error
        message:
          type: string
    StatusCertificates:
      description: This is a overview of all certificates status
      type: object
      required:
        - total
        - pending
        - ready
        - failed
        - unknown
      properties:
        total:
          type: integer
        pending:
          type: integer
        ready:
          type: integer
        failed:
          type: integer
        unknown:
          type: integer
    Certificate:
      description: This is a single certificate for consuming on clients
      type: object
      required:
        - domain
        - authority
        - key
        - crt
      properties:
        domain:
          description: The domain for which the certificate is issued
          type: string
        authority:
          description: The certificate authority for the certificate
          type: string
        key:
          description: The private key for the certificate
          type: string
        crt:
          description: The certificate for the certificate
          type: string